test: split test_pipeline_tracker.py into 4 per-concern files
Split the 883-line test_pipeline_tracker.py into per-concern files mirroring
the existing test_record_pipeline_step_complete.py split pattern:

- test_record_pipeline_step.py (~290 LoC): TestRecordPipelineStepInit,
  TestRecordPipelineStepGateClosed, TestRecordPipelineStepStatus
- test_complete_run_skill_result.py (~145 LoC): TestCompleteRunSkillResult,
  TestGetPipelineReportIncludesTrackerGaps
- test_open_kitchen_auto_init_tracker.py (~228 LoC): module helper
  _configure_open_kitchen_mock + TestOpenKitchenAutoInitTracker
- test_pipeline_tracker_authority.py (~217 LoC): TestCheckPipelineDepsImmutableTarget,
  TestSelectTrackerAuthority, TestRestoreReservedTrackerAuthority,
  TestSelectTrackerTarget

Mechanical lift: no production code changes, no behavior changes, no
test assertion changes. Each file carries pytestmark = [layer("server"),
small]. Class-level _setup autouse fixtures preserved verbatim where the
class-level setup exceeds what tests/server/conftest.py does.

Refs: #4610
